[Feat] 트레이너 회원가입 플로우 네비게이션 구현
