+++ TnT/Projects/Presentation/Sources/Onbarding/Trainer/Login/Connection/ConnectedTraineeProfileView.swift
@@ -83,12 +83,6 @@ public struct ConnectedTraineeProfileView: View {
                 }
             }
             .navigationBarBackButtonHidden()
-            .navigationDestination(
-                isPresented: Binding(get: { store.viewState.isNavigating }, set: { store.send(.setNavigating($0))})
-            ) {
-                // temp
-                // go home
-            }
         }
         .background(Color.neutral800)
     }
